Use size_t for `num_nodes` and `num_requests`

+++ day-13-cf-425-d.cpp
@@ -88,7 +88,7 @@ public:
 
 int main()
 {
-    int num_nodes, num_requests;
+    size_t num_nodes, num_requests;
 
     std::cin >> num_nodes >> num_requests;
